SummonerId changed from int to long

+++ RiotSharp/IRiotApi.cs
@@ -13,30 +13,30 @@ namespace RiotSharp
 {
     public interface IRiotApi
     {
-        Summoner GetSummoner(Region region, int summonerId);
-        Task<Summoner> GetSummonerAsync(Region region, int summonerId);
-        List<Summoner> GetSummoners(Region region, List<int> summonerIds);
-        Task<List<Summoner>> GetSummonersAsync(Region region, List<int> summonerIds);
+        Summoner GetSummoner(Region region, long summonerId);
+        Task<Summoner> GetSummonerAsync(Region region, long summonerId);
+        List<Summoner> GetSummoners(Region region, List<long> summonerIds);
+        Task<List<Summoner>> GetSummonersAsync(Region region, List<long> summonerIds);
         Summoner GetSummoner(Region region, string summonerName);
         Task<Summoner> GetSummonerAsync(Region region, string summonerName);
         List<Summoner> GetSummoners(Region region, List<string> summonerNames);
         Task<List<Summoner>> GetSummonersAsync(Region region, List<string> summonerNames);
-        SummonerBase GetSummonerName(Region region, int summonerId);
-        Task<SummonerBase> GetSummonerNameAsync(Region region, int summonerId);
-        List<SummonerBase> GetSummonersNames(Region region, List<int> summonerIds);
-        Task<List<SummonerBase>> GetSummonersNamesAsync(Region region, List<int> summonerIds);
+        SummonerBase GetSummonerName(Region region, long summonerId);
+        Task<SummonerBase> GetSummonerNameAsync(Region region, long summonerId);
+        List<SummonerBase> GetSummonersNames(Region region, List<long> summonerIds);
+        Task<List<SummonerBase>> GetSummonersNamesAsync(Region region, List<long> summonerIds);
         List<Champion> GetChampions(Region region, bool freeToPlay = false);
         Task<List<Champion>> GetChampionsAsync(Region region, bool freeToPlay = false);
         Champion GetChampion(Region region, int championId);
         Task<Champion> GetChampionAsync(Region region, int championId);
-        Dictionary<long, List<MasteryPage>> GetMasteryPages(Region region, List<int> summonerIds);
-        Task<Dictionary<long, List<MasteryPage>>> GetMasteryPagesAsync(Region region, List<int> summonerIds);
-        Dictionary<long, List<RunePage>> GetRunePages(Region region, List<int> summonerIds);
-        Task<Dictionary<long, List<RunePage>>> GetRunePagesAsync(Region region, List<int> summonerIds);
-        Dictionary<long, List<League>> GetLeagues(Region region, List<int> summonerIds);
-        Task<Dictionary<long, List<League>>> GetLeaguesAsync(Region region, List<int> summonerIds);
-        Dictionary<long, List<League>> GetEntireLeagues(Region region, List<int> summonerIds);
-        Task<Dictionary<long, List<League>>> GetEntireLeaguesAsync(Region region, List<int> summonerIds);
+        Dictionary<long, List<MasteryPage>> GetMasteryPages(Region region, List<long> summonerIds);
+        Task<Dictionary<long, List<MasteryPage>>> GetMasteryPagesAsync(Region region, List<long> summonerIds);
+        Dictionary<long, List<RunePage>> GetRunePages(Region region, List<long> summonerIds);
+        Task<Dictionary<long, List<RunePage>>> GetRunePagesAsync(Region region, List<long> summonerIds);
+        Dictionary<long, List<League>> GetLeagues(Region region, List<long> summonerIds);
+        Task<Dictionary<long, List<League>>> GetLeaguesAsync(Region region, List<long> summonerIds);
+        Dictionary<long, List<League>> GetEntireLeagues(Region region, List<long> summonerIds);
+        Task<Dictionary<long, List<League>>> GetEntireLeaguesAsync(Region region, List<long> summonerIds);
         Dictionary<string, List<League>> GetLeagues(Region region, List<string> teamIds);
         Task<Dictionary<string, List<League>>> GetLeaguesAsync(Region region, List<string> teamIds);
         Dictionary<string, List<League>> GetEntireLeagues(Region region, List<string> teamIds);
@@ -45,8 +45,8 @@ namespace RiotSharp
         Task<League> GetChallengerLeagueAsync(Region region, Queue queue);
         League GetMasterLeague(Region region, Queue queue);
         Task<League> GetMasterLeagueAsync(Region region, Queue queue);
-        Dictionary<long, List<TeamEndpoint.Team>> GetTeams(Region region, List<int> summonerIds);
-        Task<Dictionary<long, List<TeamEndpoint.Team>>> GetTeamsAsync(Region region, List<int> summonerIds);
+        Dictionary<long, List<TeamEndpoint.Team>> GetTeams(Region region, List<long> summonerIds);
+        Task<Dictionary<long, List<TeamEndpoint.Team>>> GetTeamsAsync(Region region, List<long> summonerIds);
         Dictionary<string, TeamEndpoint.Team> GetTeams(Region region, List<string> teamIds);
         Task<Dictionary<string, TeamEndpoint.Team>> GetTeamsAsync(Region region, List<string> teamIds);
         MatchDetail GetMatch(Region region, long matchId, bool includeTimeline = false);
